+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 1.0.0b2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Add renames for a few webdav interfaces.
+  In Zope 4.3 webdav will return, but until then they give real errors for zodbverify on Python 3.
+  [maurits]
 
 
 1.0.0b1 (2020-03-03)